More refactoring, added worker mesh rendering
