+++ src/ui/elements/output.ts
@@ -8,6 +8,7 @@ export class OutputElement {
 
     public constructor() {
         this._id = '_' + Math.random().toString(16);
+        this._message = new UIMessageBuilder();
     }
 
     public generateHTML() {
@@ -27,12 +28,16 @@ export class OutputElement {
         element.classList.remove('border-error');
     }
 
+    private _message: UIMessageBuilder;
+
     public setMessage(message: UIMessageBuilder, style: OutputStyle) {
         const element = document.getElementById(this._id) as HTMLDivElement;
         ASSERT(element !== null);
 
+        this._message = message;
+
         this.clearMessage();
-        element.innerHTML = message.toString();
+        element.innerHTML = this._message.toString();
         switch (style) {
             case 'success':
                 element.classList.add('border-success');
@@ -45,4 +50,35 @@ export class OutputElement {
                 break;
         }
     }
+
+    public addMessage(message: UIMessageBuilder) {
+        this._message.join(message);
+
+        const element = document.getElementById(this._id) as HTMLDivElement;
+        ASSERT(element !== null);
+
+        element.innerHTML = this._message.toString();
+        return this;
+    }
+
+    public setStyle(style: OutputStyle) {
+        const element = document.getElementById(this._id) as HTMLDivElement;
+        ASSERT(element !== null);
+
+        element.classList.remove('border-success');
+        element.classList.remove('border-warning');
+        element.classList.remove('border-error');
+
+        switch (style) {
+            case 'success':
+                element.classList.add('border-success');
+                break;
+            case 'warning':
+                element.classList.add('border-warning');
+                break;
+            case 'error':
+                element.classList.add('border-error');
+                break;
+        }
+    } 
 }
