+++ src/ui/misc.ts
@@ -5,26 +5,34 @@ export class UIMessageBuilder {
         this._messages = [];
     }
 
+    public static create() {
+        return new UIMessageBuilder();
+    }
+
     public addHeading(message: string) {
         this.addBold(message + ':');
+        return this;
     }
 
     public addBold(...messages: string[]) {
         for (const message of messages) {
             this._messages.push(`<b>${message}</b>`);
         }
+        return this;
     }
 
     public add(...messages: string[]) {
         for (const message of messages) {
             this._messages.push(message);
         }
+        return this;
     }
 
     public addItem(...messages: string[]) {
         for (const message of messages) {
             this._messages.push('• ' + message);
         }
+        return this;
     }
 
     public toString(): string {
@@ -35,5 +43,9 @@ export class UIMessageBuilder {
         const builder = new UIMessageBuilder();
         builder.addItem(string);
         return builder;
+    }
+
+    public join(builder: UIMessageBuilder) {
+        this._messages.push(...builder._messages);
     }
 }
