+++ styles.css
@@ -538,7 +538,6 @@ svg {
 
 .button-loading {
   box-shadow: 0 0 0 0 rgba(0, 0, 0, 1);
-  transform: scale(1);
   animation: pulse 2s infinite;
 }
